- Move blog tasks to separate file - Fix typo
